pyproject.toml
[project]
name = "duro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dill>=0.4.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "smolagents>=1.24.0",
    "temporalio>=1.24.0",
]

[project.optional-dependencies]
gcs = [
    "google-cloud-storage>=2.0",
]
s3 = [
    "boto3>=1.28",
]
r2 = [
    "boto3>=1.28",
]

[dependency-groups]
dev = [
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
